+++ src/KingmakerGunslinger/Acadamae/AcadamaeCastRequest.cs
@@ -3,6 +3,7 @@ namespace KingmakerGunslinger.Acadamae
     internal sealed class AcadamaeCastRequest
     {
         internal bool HasFeat { get; set; }
+        internal bool AccelerationModeActive { get; set; }
         internal bool IsRealSpell { get; set; }
         internal bool HasSpellbook { get; set; }
         internal bool IsPreparedInvocation { get; set; }